requirements.txt
PyJWT~=2.10.1
Flask~=3.1.0
pymongo~=4.11.1
eventlet~=0.39.0
Flask-Cors~=5.0.0
Flask-SocketIO~=5.5.1
python-dotenv~=1.0.1
playwright~=1.50.0
beautifulsoup4~=4.13.3
certifi~=2025.1.31
Werkzeug~=3.1.3
selenium~=4.28.1
TwoCaptcha~=0.0.1
marshmallow~=3.26.1
binance~=0.3
requests
selenium_stealth
python-binance
webdriver-manager~=4.0.2
